requirements.txt
termcolor
markdown
pyyaml
typing_extensions
matrix-nio[e2e]
jinja2
requests